Revert "More attempts at fixing the notifier bug."
This reverts commit 9897764aa52fbd31530d89ba54e7bd37cb9ffb60.

+++ WPILib/Utility.cs
@@ -36,10 +36,10 @@ namespace WPILib
         /// <summary>
         /// Read the microsecond timer from the FPGA
         /// </summary>
-        public static uint GetFPGATime()
+        public static long GetFPGATime()
         {
             int status = 0;
-            uint value = HAL.GetFPGATime(ref status);
+            long value = HAL.GetFPGATime(ref status);
             return value;
         }
